requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
plotly>=5.10.0
folium>=0.14.0
streamlit-folium>=0.11.0
requests>=2.28.0
fpdf2>=2.7.0
